requirements.txt
# 深度学习框架
torch==2.3.1
transformers==4.44.2
tokenizers==0.19.1

# Web界面
gradio==4.44.1
fastapi==0.115.4
uvicorn==0.32.0

# 机器学习工具
scikit-learn==0.24.1
shap==0.41.0
lightgbm==4.4.0
xgboost==0.90

# 数据处理
numpy==1.22.0
pandas==2.0.3
scipy==1.10.1

# 可视化
matplotlib==3.5.2
seaborn==0.11.2

# NLP工具
jieba==0.42.1
regex==2024.5.15

# 图像处理
Pillow>=9.2.0
opencv-contrib-python>=4.10.0.84

# 工具库
tqdm==4.66.4
psutil==5.8.0
python-multipart==0.0.17

# 日志和监控
tensorboard==2.14.0
tensorboardX==2.5.1

# 开发工具
black==22.1.0
ruff==0.7.1
mypy>=0.9.0

# 测试工具
pytest==6.0.0
pytest-cov==2.0.0

# 文件处理
aiofiles==23.2.1
fsspec==2024.6.0
